Extract a function that determine a subcommand
