+++ src/main.rs
@@ -2,5 +2,6 @@ extern crate miteras;
 use miteras::cli;
 
 fn main() {
-    cli::run();
+    let matches = cli::build_app().get_matches();
+    cli::run(matches);
 }
