+++ tests/main.rs
@@ -1,4 +1,6 @@
 extern crate assert_cmd;
+#[macro_use]
+extern crate clap;
 extern crate predicates;
 
 use assert_cmd::prelude::*;
@@ -22,4 +24,13 @@ fn with_help_args() {
         .assert()
         .success()
         .stdout(predicate::str::contains(CLI_DESCRIBE));
+}
+
+#[test]
+fn with_version_args() {
+    let mut cmd = Command::cargo_bin("miteras").unwrap();
+    cmd.arg("--version")
+        .assert()
+        .success()
+        .stdout(format!("miteras {}\n", crate_version!()));
 }
